Modernise package: parquet I/O, test suite, multiprocessing fixes, doc updates
Dependency / build
- Replace pickle-based pool map and block design files with parquet (pyarrow)
- Add pyarrow ^23 as a runtime dependency

Bug fixes
- imputation.py: fix pattern-mask conversion (astype(str) → astype(int).astype(bool))
  so that NearestNeighbors receives a non-empty feature array
- simulations.py: remove dead groupby.mean() call that raised TypeError under pandas 3

Multiprocessing (batch_fx.py)
- Remove hardcoded time.sleep(40) after parallel permutation runs
- Remove unnecessary Manager / managed list / Lock (imap_unordered delivers results
  directly; no shared state is needed)
- Remove redundant pool.close() / pool.join() after the with-Pool context manager
- Unify serial and parallel code paths; drop unused imports (time, current_process, Manager)

Tests
- Add tests/ with conftest.py and four test modules covering all public classes
  (44 tests, all passing)

Documentation
- Update README: parquet file format for pool map and block design, correct dependency
  list (sklearn → scikit-learn, add pyarrow), remove outdated macOS multiprocessing
  caveat, remove completed TO DO item
- Update docstrings in batch_fx, old_fashioned, simulations to reflect parquet I/O
  and corrected multiprocessing description

+++ LIMBR/imputation.py
@@ -15,7 +15,6 @@ from sklearn.neighbors import NearestNeighbors
 import math
 import json
 from ctypes import c_int
-import pickle
 from multiprocess import Pool, current_process, Manager
 from functools import partial
 from sklearn import preprocessing
@@ -154,7 +153,7 @@ missing values with corresponding averages.
             """
 
             #drop missing columns given missingness pattern
-            newarr = comparr[:,~np.array(list(pattern)).astype(bool)]
+            newarr = comparr[:,~np.array(list(pattern)).astype(int).astype(bool)]
             #fit nearest neighbors
             nbrs = NearestNeighbors(n_neighbors=self.NN).fit(newarr)
             outa = []
@@ -162,7 +161,7 @@ missing values with corresponding averages.
             for rowind, row in enumerate(origarr[inds]):
                 outl = []
                 #get indexes of given rows nearest neighbors
-                indexes = nbrs.kneighbors([origarr[inds[rowind],~np.array(list(pattern)).astype(bool)]],return_distance=False)
+                indexes = nbrs.kneighbors([origarr[inds[rowind],~np.array(list(pattern)).astype(int).astype(bool)]],return_distance=False)
                 #get array of nearest neighbors
                 means = np.mean(comparr[indexes[0][1:]], axis=0)
                 #iterate over entries in each row
